Code/Intelligence_Reservation_Agent (Project-Code)/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall==0.1.0
certifi==2018.11.29
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
decorator==4.3.0
django==3.1.2
Faker==4.14.0
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.22.4
google-auth==1.22.1
googleapis-common-protos==1.52.0
ipykernel==5.1.4
ipython==7.12.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
mkl-service==2.3.0
matplotlib==3.2.0
numpy==1.18.4
numpydoc==0.9.2
oauthlib==3.1.0
pandas==1.0.1
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2020.8.13
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
pywin32==225
pywin32-ctypes==0.2.0
pyzmq==18.1.1
requests==2.21.0
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
srsly==1.0.0
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
tqdm==4.45.0
urllib3==1.24.1
virtualenv==20.0.33
virtualenv-clone==0.5.4
wcwidth==0.1.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.12.0
zipp==2.2.0
